+++ exec/Kathu/App/Tools/EventQueue.hs
@@ -1,32 +1,27 @@
-{-# LANGUAGE OverloadedStrings   #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-
 module Kathu.App.Tools.EventQueue where
 
 import           Apecs
-import           Apecs.Physics
 import           Control.Concurrent.MVar
-import           Control.Monad               (forM_, mapM_, when)
+import           Control.Monad               (mapM_)
 import           Control.Lens
 import qualified Data.Map                    as Map
-import           Linear.V2                   (V2(..))
 
 import           Kathu.App.Data.Library
 import           Kathu.App.Graphics.Image
 import           Kathu.App.System
 import           Kathu.App.Tools.Commands
 import           Kathu.App.Tools.ToolMode
+import           Kathu.App.Tools.ToolSystem
 import           Kathu.App.World             (loadWorldSpace)
-import           Kathu.Entity.Components     (Local, simpleIdentity)
-import           Kathu.Graphics.Camera
-import           Kathu.Util.Apecs
+import           Kathu.Entity.System         (Debug(..))
 import           Kathu.World.WorldSpace      (WorldSpace, emptyWorldSpace, worldID)
-    
+
 -- Events that the app receives
 data AppEvent
     = UseToolMode ToolMode
     | LoadWorldSpace (WorldSpace ImageID)
     | NewWorldSpace
+    | ToggleDebug
     | RunSystem (SystemT' IO ())
 
 -- Events that the editor receives
@@ -47,32 +42,13 @@ handleEvents queue  _ = do
 
 handleEvent :: AppEvent -> SystemT' IO ()
 handleEvent event = case event of
-    RunSystem action -> action
-    UseToolMode newMode -> do
-        prevMode <- get global
-        global   $= newMode
-        -- the new mode is a free cam mode, but not the last
-        -- let's create a new free cam entity, and give it the camera
-        when (usesFreeCam newMode && not (usesFreeCam prevMode)) $ do
-            playerEty  <- getUnique
-            cmap $ \(_ :: Camera) -> Not :: Not Camera
-
-            freeCamEty <- newEntity (simpleIdentity "ToolMode Free Camera", DynamicBody)
-            case playerEty of
-                Just (cam :: Camera, p :: Position) -> freeCamEty $= (cam, p)
-                Nothing                             -> freeCamEty $= (defaultCamera, Position (V2 0 0))
-        -- the new mode is not a free cam mode, but the last is
-        -- let's destroy the free cam entity, and give the camera back to the player
-        when (usesFreeCam prevMode && not (usesFreeCam newMode)) $ do
-            lastCamEty <- getUnique
-            playerEty  <- getUnique
-            -- Removes camera, then destroys free cam entity
-            forM_ lastCamEty $ \(_ :: Camera, ety) -> do
-                cmap $ \(_ :: Camera) -> Not :: Not Camera
-                destroyEntity ety
-            -- Gives the camera to the player
-            forM_ playerEty $ \(_ :: Local, ety) ->
-                ety $= defaultCamera
+    RunSystem action ->
+        action
+    ToggleDebug -> do
+        Debug isDebug <- get global
+        global        $= Debug (not isDebug)
+    UseToolMode newMode ->
+        handleUseToolModeEvent newMode
     NewWorldSpace ->
         loadWorldSpace emptyWorldSpace
     LoadWorldSpace worldspace -> do